feat: added a new page 'manual'
